+++ src/main/java/com/reis/HotelManagementSystem_APi/dto/StaySummaryDTO.java
@@ -1,5 +1,6 @@
 package com.reis.HotelManagementSystem_APi.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.reis.HotelManagementSystem_APi.entities.Stay;
@@ -9,7 +10,7 @@ public class StaySummaryDTO {
 	private Long id;
 	private LocalDateTime checkInDate;
 	private LocalDateTime checkOutDate;
-	private Double totalValue;
+	private BigDecimal totalValue;
 		
 	private GuestSummaryDTO guestSummaryDTO;
 	
@@ -36,7 +37,7 @@ public class StaySummaryDTO {
 		return checkOutDate;
 	}
 
-	public Double getTotalValue() {
+	public BigDecimal getTotalValue() {
 		return totalValue;
 	}
